fix(cli): resolve NameError for show_banner in cli_network (#323)
Move show_banner() to cli_shared.py to avoid circular import between
cli.py and cli_network.py. Both modules now import it from the shared
module, fixing the crash during network startup/shutdown.
