+++ src/openagents/client/cli_shared.py
@@ -9,6 +9,7 @@ import sys
 
 import typer
 from rich.console import Console
+from rich.panel import Panel
 
 # Ensure UTF-8 output on Windows (avoids charmap codec errors with emoji)
 if sys.platform == "win32" and not os.environ.get("PYTHONIOENCODING"):
@@ -38,3 +39,23 @@ VERBOSE_MODE = False
 OPENAGENTS_API_BASE = "https://endpoint.openagents.org/v1"
 OPENAGENTS_RELAY_URL = "wss://relay.openagents.org"
 LOCALHOST_HOSTS = ["localhost", "127.0.0.1", "0.0.0.0", "::1", "local"]
+
+
+def show_banner():
+    """Show a beautiful startup banner"""
+    banner_text = """
+[bold blue]   ___                              ___                          _       [/bold blue]
+[bold blue]  / _ \\ _ __    ___  _ __           /   \\  __ _   ___  _ __   | |_  ___ [/bold blue]
+[bold blue] | | | | '_ \\  / _ \\| '_ \\         / /\\ / / _` | / _ \\| '_ \\  | __|/ __[/bold blue]
+[bold blue] | |_| | |_) ||  __/| | | |       / /_// | (_| ||  __/| | | | | |_\\__ \\ [/bold blue]
+[bold blue]  \\___/| .__/  \\___||_| |_|      /___,'   \\__, | \\___||_| |_|  \\__|___/[/bold blue]
+[bold blue]       |_|                              |___/                        [/bold blue]
+
+[bold cyan]AI Agent Networks for Open Collaboration[/bold cyan]
+[dim]   Create and manage distributed AI agent networks with ease[/dim]
+"""
+    console.print(Panel(
+        banner_text.strip(),
+        border_style="blue",
+        expand=False
+    ))
